Update fr.xml
Update in French. I changed some sentences to one word as "école ou lycée ou université" as "cours".
This meaning is "at school or college or university" as a meaning in French in short sentence "cours". that's why, I updated that too. It's clarify the long way to say it.  :)